cli: Move commands to their own files
To make the previously monolithic Main class more readable.
