+++ detekt.yml
@@ -38,6 +38,7 @@ style:
     excludeImports: '
       ch.frankel.slf4k.*,
       com.here.ort.analyzer.managers.*,
+      com.here.ort.commands.*,
       com.here.ort.downloader.vcs.*,
       com.here.ort.reporter.reporters.*,
       com.here.ort.scanner.scanners.*
